+++ src/tests/Unpacked.test.ts
@@ -1,17 +1,18 @@
-import type { Unpacked } from "../utilityTypes/Unpacked";
+import type { Unpacked } from "../utilityTypes/Unpacked.ts";
 
-describe("Unpacked", () => {
-    it("will extract a simple array type", () => {
+import { assertEquals } from "jsr:@std/assert";
+Deno.test("Unpacked", () => {
+    Deno.test("will extract a simple array type", () => {
         const s: Unpacked<Array<string>> = "";
-        expect(s).toBe("");
+        assertEquals(s, "");
     });
-    it("will extract a complex array type", () => {
-        const complexArray = [{ type: "demo", value: "foobar"}];
+    Deno.test("will extract a complex array type", () => {
+        const complexArray = [{ type: "demo", value: "foobar" }];
         const s: Unpacked<typeof complexArray> = complexArray[0];
-        expect(s.value).toBe("foobar");
+        assertEquals(s.value, "foobar");
     });
-    it("will return the type for non-array types", () => {
+    Deno.test("will return the type for non-array types", () => {
         const s: Unpacked<string> = "";
-        expect(s).toBe("");
+        assertEquals(s, "");
     });
 });