+++ src/tests/async.test.ts
@@ -1,55 +1,68 @@
 import { Duration } from "luxon";
 
-import { settleMap, wait } from "../async";
-import { seconds } from "../time";
+import { settleMap, timeout, wait } from "../async.ts";
+import { seconds } from "../time.ts";
 
-describe("settledMap", () => {
-    it("will async map over array", async () => {
-        const s = await settleMap(async a => Promise.resolve(a + 1), [2, 4, 6]);
-        expect(s).toHaveLength(3);
-        const t = s.map(s => (s as PromiseFulfilledResult<number>).value);
-        expect(t).toEqual([3, 5, 7]);
+import { assert, assertEquals, assertRejects } from "jsr:@std/assert";
+
+Deno.test("settledMap", () => {
+    Deno.test("will async map over array", async () => {
+        const s = await settleMap(async (a) => await Promise.resolve(a + 1), [
+            2,
+            4,
+            6,
+        ]);
+        assertEquals(s.length, 3);
+        const t = s.map((s) => (s as PromiseFulfilledResult<number>).value);
+        assertEquals(t, [3, 5, 7]);
     });
-    it("will async map over array and reject correctly", async () => {
-        const s = await settleMap(async () => Promise.reject("foo error"), [2, 4, 6]);
-        expect(s).toHaveLength(3);
-        const t = s.map(s => (s as PromiseRejectedResult).reason);
-        expect(t).toEqual(["foo error", "foo error", "foo error"]);
+    Deno.test("will async map over array and reject correctly", async () => {
+        const s = await settleMap(async () => await Promise.reject("foo error"), [
+            2,
+            4,
+            6,
+        ]);
+        assertEquals(s.length, 3);
+        const t = s.map((s) => (s as PromiseRejectedResult).reason);
+        assertEquals(t, ["foo error", "foo error", "foo error"]);
     });
 });
 
-describe("wait", () => {
-    it("will wait for an amount of time", async () => {
+Deno.test("wait", () => {
+    Deno.test("will wait for an amount of time", async () => {
         const start = Date.now();
         await wait(300);
         const end = Date.now();
-        expect(end-start).toBeGreaterThan(299);
-        expect(end-start).toBeLessThan(310);
+        assert(end - start > 299);
+        assert(end - start < 310);
     });
-    it("should also work with a duration", async () => {
+    Deno.test("should also work with a duration", async () => {
         const start = Date.now();
         await wait(seconds(1));
         const end = Date.now();
-        expect(end-start).toBeGreaterThan(999);
-        expect(end-start).toBeLessThan(1150);
+        assert(end - start > 999);
+        assert(end - start < 1150);
     });
 });
 
-describe("timeout", () => {
-    it("will trigger when the timeout is reached", async () => {
+Deno.test("timeout", () => {
+    Deno.test("will trigger when the timeout is reached", async () => {
         const waitingPromise = wait(1000);
-        const tu = Promise.timeout(waitingPromise, 500);
-        expect(tu).rejects.toBeTruthy();
+        const tu = await timeout(waitingPromise, 500);
+        assert(tu);
         await waitingPromise;
     });
-    it("will not trigger when the timeout is not reached", async () => {
-        const tu = Promise.timeout(Promise.resolve(true), 500);
-        expect(tu).resolves.toBeTruthy();
+    Deno.test("will not trigger when the timeout is not reached", async () => {
+        const tu = await timeout(Promise.resolve(true), 500);
+        assert(tu);
     });
-    it("will also work with a duration", async () => {
+    Deno.test("will also work with a duration", async () => {
         const waitingPromise = wait(1000);
-        const tu = Promise.timeout(waitingPromise, Duration.fromObject({milliseconds: 500}));
-        expect(tu).rejects.toBeTruthy();
+        const tu = timeout(
+            waitingPromise,
+            Duration.fromObject({ milliseconds: 500 }),
+        );
+        assertRejects(() => tu);
         await waitingPromise;
     });
 });