+++ src/tests/index.test.ts
@@ -1,3 +1,0 @@
-describe("Test index", () => {
-    it("", () => { expect(2).toBe(2); });
-});