+++ src/tests/unit.test.ts
@@ -1,14 +1,16 @@
-import type { Unit} from "../utilityTypes/Unit";
-import { unit } from "../utilityTypes/Unit";
+import type { Unit } from "../utilityTypes/Unit.ts";
+import { unit } from "../utilityTypes/Unit.ts";
 
-describe("Unit type", () => {
-    it("can be returned by a function", () => {
+import { assert, assertEquals } from "jsr:@std/assert";
+
+Deno.test("Unit type", () => {
+    Deno.test("can be returned by a function", () => {
         const f = (): Unit => unit();
-        expect(f()).toBeTruthy();
-        expect(f()).toBe(unit());
+        assert(f());
+        assertEquals(f(), unit());
     });
-    it("it equals a truthy expression", () => {
-        expect(unit()).toBeTruthy();
-        expect(1 == 1).toEqual(unit());
+    Deno.test("it equals a truthy expression", () => {
+        assert(unit());
+        assertEquals(1 == 1, unit());
     });
 });