+++ src/utilityTypes/Reveal.ts
@@ -1,7 +1,7 @@
 /**
  * Reveal all properties of a deeply-inherited type.
- * 
+ *
  * Use this for (static) debugging purposes. Avoid using it in production.
  */
-// eslint-disable-next-line @typescript-eslint/ban-types
+// deno-lint-ignore ban-types
 export type Reveal<T> = { [K in keyof T]: T[K] } & {};