+++ src/utilityTypes/Unit.ts
@@ -1,10 +1,13 @@
-import type { Nominal } from "./Nominal";
+import * as z from "zod";
 
-/** 
- * Unit type for functional programming. For simplicity and minimal space consumption, this maps to `true` 
+const unitBrand: z.ZodBranded<z.ZodLiteral<true>, "unit"> = z.literal(true)
+    .brand("unit");
+
+/**
+ * Unit type for functional programming. For simplicity and minimal space consumption, this maps to `true`
  * @important Note that this is different to `void` which basically is undefined and therefor *falsy*
  */
-export type Unit = Nominal<true, "unit">;
+export type Unit = z.infer<typeof unitBrand>;
 
 /** Create a value of type Unit */
 export const unit = (): Unit => true as Unit;